requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.8.3
cffi==1.17.1
cryptography==45.0.4
greenlet==3.2.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jeepney==0.9.0
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.7.0
mysql-connector-python==9.3.0
prompt_toolkit==3.0.51
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.0
rich==14.0.0
SecretStorage==3.3.3
sentry-sdk==2.34.1
simple-term-menu==1.6.6
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
wcwidth==0.2.13
